Fix for the bad access error.
Adding a copy constructor and assignment operator to the BatchCommand and CustomCommand structs.

Location of the the bug:

File: CCLabel.h
Line: 1794
_batchCommands.resize(_batchNodes.size());

When a batch command vector of size > 0 is being resized to a larger vector it will copy the data to a new object and then destroy the old one. This causes BatchCommand to release object pointers in its destructor after those pointer have been copied over to a new object. Same issue applies to the customCommand struct because it also releases objects in its destructor, and does not have a copy or assignment constructor.